Fix Codeforces solutions

+++ simplex/testing/Codeforces.Gym.104506/B. Simple Problem/pre.cpp
@@ -8,6 +8,6 @@
 using namespace std;
 using vi = vector<int>;
 using vvi = vector<vi>;
-using db = double;
+using db = long double;
 using vf = vector<db>;
 using vvf = vector<vf>;
